fix state managment in Edit Dialog
should work in theory. There are some things acting a bit
weird for example for structs, but i really think, that might be
an unrelated problem with the backend. Because i saw some crashes
there don't!

Oh and there is still some decoration missing the names of the od
fields are not there yet.

The NumberField and StringFields should now be independent from the
type, which is hopefully an improvement =^).