Gentlemen we got him!
Holy shit this took a while, you might have noticed that the graphs
sometimes just don't work properly and then you have to refresh them and
then they kind of work again, but at some point if the values come in
very high frequencies this fails and fixing it took 6h hours
"YEAH KILL ME".

It was a deadlock with a bounded channel of capacity 10. Where 12 frames
had to arrive more or less at the same time, which will lead to
blocking on the channel. However the consumer side couldn't continue,
because the store lock got held over the sending on the channel.

We sometimes had the graphs acting weird on socketcan, which is
basically a slightly less critical deadlock, wherethe capacity was just
really full, which lead to observers not beeing clean up properly. Let's
go memory leaks in rust; There is a certain skill involved here.
After moving to TCP this really became a problem because TCP tends to
burst a lot of CAN frames together, which lead to more then 12 values for
a object entry beeing received together, which then lead to the deadlock
happening and killing the backend completely, because the main CNL lock
gets held, which leads to the backend not responding to anything.

Just a note maybe i'm just mad, but why doesn't rust destruct values
early, sometimes it's sooo annoying especially with mutexes.

To understand the fix drop(store) destructs the mutex, which releases it
(RAII), without this it would be released at the end of the function.